+++ include/MeshKernel/Polygons.hpp
@@ -55,7 +55,7 @@ namespace meshkernel
         /// @param[in] endIndex The end index
         /// @param[in] refinementDistance The chosen refinement distance
         /// @return refinedPolygon The computed polygon
-        std::vector<Point> RefineFirstPolygon(int startIndex, int endIndex, double refinementDistance) const;
+        std::vector<Point> RefineFirstPolygon(size_t startIndex, size_t endIndex, double refinementDistance) const;
 
         /// @brief Makes a new polygon from an existing one, by offsetting it by a distance (copypol)
         /// @param[in] distance The offset distance
@@ -67,7 +67,7 @@ namespace meshkernel
         /// @param[in] point The point to check
         /// @param[in] polygonIndex The index of the polygon to account for
         /// @return True if it is included, false otherwise
-        bool IsPointInPolygon(Point point, int polygonIndex) const;
+        bool IsPointInPolygon(Point point, size_t polygonIndex) const;
 
         /// @brief Checks if a point is included in any of the polygons (dbpinpol_optinside_perpol)
         /// @param[in] point The point to check
